src/main/java/edu/hw6/task4/OutputStreamCompositionManager.java
package edu.hw6.task4;

import java.io.BufferedOutputStream;
import java.io.IOException;
import java.io.OutputStream;
import java.io.OutputStreamWriter;
import java.io.PrintWriter;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Path;
import java.util.zip.CRC32;
import java.util.zip.CheckedOutputStream;

public final class OutputStreamCompositionManager {
    private OutputStreamCompositionManager() {}

    public static void writeToFile(Path filePath, String message) {
        try (
            OutputStream outputStream = Files.newOutputStream(filePath);
            CheckedOutputStream checkedOutputStream = new CheckedOutputStream(outputStream, new CRC32());
            BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(checkedOutputStream);
            OutputStreamWriter outputStreamWriter =
                new OutputStreamWriter(bufferedOutputStream, StandardCharsets.UTF_8);
            PrintWriter printWriter = new PrintWriter(outputStreamWriter)
        ) {
            printWriter.println(message);
        } catch (IOException e) {
            throw new RuntimeException(e);
        }
    }
}
